role check on editing/ deleting

+++ controllers/ProjectController.js
@@ -27,25 +27,31 @@ export const getProject = async (req, res) => {
 }
 
 export const createProject = async (req, res) => {
-    const { projectTitle, startDate, author } = req.body;
+    const { projectTitle, startDate, author, role } = req.body;
 
-    const newProject = new ProjectModel({ projectTitle, startDate, author })
+    if(role !== process.env.NODE_ENV_ADMIN_SECRET || role !== process.env.NODE_ENV_USER_SECRET ){
+      res.json("You do not have permission");
+    } else {
+        const newProject = new ProjectModel({ projectTitle, startDate, author })
+        try {
+            await newProject.save();
 
-    try {
-        await newProject.save();
-
-        res.status(201).json("Project Created");
-    } catch (error) {
-        res.status(409).json({ message: error.message });
+            res.status(201).json("Project Created");
+        } catch (error) {
+            res.status(409).json({ message: error.message });
+        }
     }
 }
 
 export const deleteProject = async (req, res) => {
     const { projectId } = req.params;
-
+    const { role } = req.body;
     if (!mongoose.Types.ObjectId.isValid(projectId)) return res.status(404).send(`No project with id: ${projectId}`);
 
-    await ProjectModel.findByIdAndRemove(projectId);
-
-    res.json("Project Deleted");
+    if(role !== process.env.NODE_ENV_ADMIN_SECRET || role !== process.env.NODE_ENV_USER_SECRET ){
+      res.json("You do not have permission");
+    } else {
+        await ProjectModel.findByIdAndRemove(projectId);
+        res.json("Project Deleted");
+    }
 }
